rename android app to RN Boilerplate
