+++ android/app/src/dev/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Gooner007 Expense Dev</string>
+    <string name="app_name">RN Boilerplate Dev</string>
 </resources>
