+++ android/app/src/main/java/com/gooner007/expense/MainActivity.java
@@ -1,4 +1,4 @@
-package com.gooner007.expense;
+package com.gooner007.reactnativeboilerplate;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
@@ -14,7 +14,7 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "Gooner007Expense";
+    return "ReactNativeBoilerplate";
   }
 
   /**
