+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Gooner007 Expense</string>
+    <string name="app_name">RN Boilerplate</string>
 </resources>
