+++ android/app/src/uat/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Gooner007 Expense UAT</string>
+    <string name="app_name">RN Boilerplate UAT</string>
 </resources>
