+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "Gooner007Expense",
-  "displayName": "Gooner007Expense"
+  "name": "ReactNativeBoilerplate",
+  "displayName": "ReactNativeBoilerplate"
 }
