+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "Gooner007Expense",
+  "name": "ReactNativeBoilerplate",
   "version": "0.0.1",
   "private": true,
   "scripts": {
